+++ common/src/main/java/net/phantompig/soy/block/SoyBlocks.java
@@ -17,10 +17,11 @@ public class SoyBlocks {
     public static final RegistrySupplier<Block> DEEPSLATE_ICEBURST_ORE = BLOCKS.register("deepslate_iceburst_ore", () -> new IceburstBlock(BlockBehaviour.Properties.of().mapColor(MapColor.COLOR_GRAY).requiresCorrectToolForDrops().sound(SoundType.DEEPSLATE).strength(4.3f, 6), 5));
 
     public static final RegistrySupplier<Block> IRON_BAMBOO = BLOCKS.register("iron_bamboo", () -> new IronBambooBlock(BlockBehaviour.Properties.of().forceSolidOn().mapColor(MapColor.COLOR_GRAY).offsetType(BlockBehaviour.OffsetType.XZ).noOcclusion().dynamicShape().sound(SoundType.METAL).strength(1.5f, 0.5f).pushReaction(PushReaction.DESTROY).isRedstoneConductor((e, g, p) -> true)));
+    public static final RegistrySupplier<Block> PACKED_IRON_BAMBOO = BLOCKS.register("packed_iron_bamboo", () -> new PackedIronBambooBlock(BlockBehaviour.Properties.of().mapColor(MapColor.COLOR_GRAY).requiresCorrectToolForDrops().sound(SoundType.METAL).strength(24, 14)));
 
     public static final RegistrySupplier<Block> COMPRESSION_TABLE = BLOCKS.register("compression_table", () -> new CompressionTableBlock(BlockBehaviour.Properties.of().mapColor(MapColor.COLOR_BROWN).sound(SoundType.WOOD).strength(2f, 1f)));
 
-    public static final RegistrySupplier<Block> HARDENING_BLOCK = BLOCKS.register("hardening_block", () -> new Block(BlockBehaviour.Properties.of().mapColor(MapColor.COLOR_LIGHT_BLUE).sound(SoundType.AMETHYST).strength(14, 24).requiresCorrectToolForDrops()));
+    public static final RegistrySupplier<Block> HARDENING_BLOCK = BLOCKS.register("hardening_block", () -> new Block(BlockBehaviour.Properties.of().mapColor(MapColor.COLOR_LIGHT_BLUE).sound(SoundType.AMETHYST).strength(40, 24).requiresCorrectToolForDrops()));
 
     public static void init() {
         BLOCKS.register();