+++ common/src/main/java/net/phantompig/soy/item/SoyItems.java
@@ -39,6 +39,7 @@ public class SoyItems {
 
     public static final RegistrySupplier<Item> IRON_BAMBOO = ITEMS.register("iron_bamboo", () -> new Item(new Item.Properties()));
     public static final RegistrySupplier<Item> IRON_BAMBOO_LEAF = ITEMS.register("iron_bamboo_leaf", () -> new Item(new Item.Properties()));
+    public static final RegistrySupplier<Item> PACKED_IRON_BAMBOO = ITEMS.register("packed_iron_bamboo", () -> new BlockItem(SoyBlocks.PACKED_IRON_BAMBOO.get(), new Item.Properties()));
 
     public static final RegistrySupplier<Item> IRON_BAMBOO_SWORD = ITEMS.register("iron_bamboo_sword", () -> new SwordItem(ToolTiers.IRON_BAMBOO, 3, -1.5f, new Item.Properties()));
     public static final RegistrySupplier<Item> IRON_BAMBOO_DAGGER = ITEMS.register("iron_bamboo_dagger", () -> new SwordItem(ToolTiers.IRON_BAMBOO, 1, -0.5f, new Item.Properties()));
@@ -156,6 +157,7 @@ public class SoyItems {
             entries.add(ULTRAHARD_STEEL_INGOT.get());
             entries.add(IRON_BAMBOO.get());
             entries.add(IRON_BAMBOO_LEAF.get());
+            entries.add(PACKED_IRON_BAMBOO.get());
             entries.add(IRON_BAMBOO_SWORD.get());
             entries.add(IRON_BAMBOO_DAGGER.get());
             entries.add(IRON_BAMBOO_RING.get());